Basic annotations - floor

+++ python_variable_annotations/2-floor.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
 import math
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     return math.floor(n)
